// packages/request-hooks/src/request/models/GarnerRecord.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * 收藏记录
 */
export type GarnerRecord = {
  /**
   * 收藏记录ID
   */
  garnerRecordId: string;
  /**
   * 创建时间
   */
  createdAt: string;
  /**
   * 更新时间
   */
  updatedAt: string;
  /**
   * 用户信息ID
   */
  userInformationId: string;
  /**
   * 公司信息ID
   */
  companyInformationId: string;
  /**
   * 职位信息ID
   */
  positionInformationId: string;
};